requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
discord==2.2.3
discord.py==2.2.3
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
filelock==3.12.0
Flask==2.3.2
frozenlist==1.3.3
gyp==0.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
oauthlib==3.2.0
pipenv==2023.6.2
platformdirs==3.5.1
psycopg==3.1.9
psycopg-binary==3.1.9
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==5.4.1
rich==13.4.1
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
tomli==2.0.1
twitchAPI==3.11.0
typing_extensions==4.6.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
virtualenv==20.23.0
virtualenv-clone==0.5.7
wadllib==1.3.6
Werkzeug==2.3.4
yapf==0.33.0
yarl==1.9.2
zipp==1.0.0
